requirements.txt
discord.py>=2.3
aiohttp>=3.9
aiosqlite>=0.19
cryptography>=41.0
APScheduler>=3.10
python-dotenv>=1.0
psutil>=5.9
fastapi>=0.110
uvicorn>=0.27
